--- src/main/java/com/matzip/thread/common/exception/ExceptionController.java
@@ -1,21 +1,0 @@
-package com.matzip.thread.common.exception;
-
-import com.matzip.thread.common.model.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class ExceptionController {
-
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public ErrorResponse invalidRequestHandler(MethodArgumentNotValidException e) {
-        ErrorResponse response = new ErrorResponse(HttpStatus.BAD_REQUEST, "bad request");
-        e.getFieldErrors().iterator().forEachRemaining(response::addFieldError);
-
-        return response;
-    }
-}
